Running & Listing Simulations

Running & Listing Simulations

Preparing Simulation Runs

It is very often the case that you want to run "batch simulations", i.e. just submit a bunch of different simulations, all using same algorithms and code but just different parameters. This scenario always requires the user to prepare a set of simulation parameter containers which are then passed into some kind of "main" function that starts the simulation.

To make the preparation part simpler we provide the following functionality:

DrWatson.dict_list — Function.

dict_list(c)

Expand the dictionary c into a vector of dictionaries. Each entry has a unique combination from the product of the Vector values of the dictionary while the non-Vector values are kept constant for all possibilities. The keys of the entries are the same.

Whether the values of c are iterable or not is of no concern; the function considers as "iterable" only subtypes of Vector.

Use the function dict_list_count to get an estimate of how many dictionaries will dict_list produce.

Examples

julia> c = Dict(:a => [1, 2], :b => 4);

julia> dict_list(c)
3-element Array{Dict{Symbol,Int64},1}:
 Dict(:a=>1,:b=>4)
 Dict(:a=>2,:b=>4)

julia> c[:model] = "linear"; c[:mode] = ["bi", "tri"];

julia> dict_list(c)
4-element Array{Dict{Symbol,Any},1}:
 Dict(:a=>1,:b=>4,:mode=>"bi",:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"bi",:model=>"linear")
 Dict(:a=>1,:b=>4,:mode=>"tri",:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"tri",:model=>"linear")

julia> c[:e] = [[1, 2], [3, 5]];

julia> dict_list(c)
8-element Array{Dict{Symbol,Any},1}:
 Dict(:a=>1,:b=>4,:mode=>"bi",:e=>[1, 2],:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"bi",:e=>[1, 2],:model=>"linear")
 Dict(:a=>1,:b=>4,:mode=>"tri",:e=>[1, 2],:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"tri",:e=>[1, 2],:model=>"linear")
 Dict(:a=>1,:b=>4,:mode=>"bi",:e=>[3, 5],:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"bi",:e=>[3, 5],:model=>"linear")
 Dict(:a=>1,:b=>4,:mode=>"tri",:e=>[3, 5],:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"tri",:e=>[3, 5],:model=>"linear")

DrWatson.dict_list_count — Function.

dict_list_count(c) -> N

Return the number of dictionaries that will be created by calling dict_list(c).

Using the above function means that you can write your "preparation" step into a single dictionary and then let it automatically expand into many parameter containers. This keeps the code cleaner but also consistent, provided that it follows one rule: Anything that is a Vector has many parameters, otherwise it is one parameter. dict_list considers this true irrespectively of what the Vector contains. This allows users to use any iterable custom type as a single "parameter" of a simulation.

See the Real World Examples for a very convenient application!

Simulation Tables

WIP. (Adding simulation runs to a table/csv/dataframe)