Naming & Saving Simulations

Naming & Saving Simulations

This page discusses numerous tools that make life easier for handling simulations. Most (if not all) of these tools are also used in the examples demonstrated in the Real World Examples page. After reading the proper documentation here it might be worth it to have a look there as well!

Naming Schemes

A robust naming scheme allows you to create quick names for simulations, create lists of simulations, check existing simulations, etc. More importantly it allows you to easily read and write simulations using a consistent naming scheme.

DrWatson.savename — Function.

savename([prefix,], c [, suffix]; kwargs...)

Create a shorthand name, commonly used for saving a file, based on the parameters in the container c (Dict, NamedTuple or any other Julia composite type, e.g. created with Parameters.jl). If provided use the prefix and end the name with .suffix (i.e. you don't have to include the . in your suffix).

The function chains keys and values into a string of the form:

key1=val1_key2=val2_key3=val3

while the keys are always sorted alphabetically. If you provide the prefix/suffix the function will do:

prefix_key1=val1_key2=val2_key3=val3.suffix

assuming you chose the default connector, see below. Notice that prefix can be any path and in addition if it ends as a path (/ or \) then the connector is ommited.

savename can be very conveniently combined with @dict or @ntuple.

Keywords

allowedtypes = default_allowed(c) : Only values of type subtyping anything in allowedtypes are used in the name. By default this is (Real, String, Symbol).

accesses = allaccess(c) : You can also specify which specific keys you want to use with the keyword accesses. By default this is all possible keys c can be accessed with, see allaccess.

digits = 3 : Floating point values are rounded to digits. In addition if the following holds:
round(val; digits = digits) == round(Int, val)
then the integer value is used in the name instead.

connector = "_" : string used to connect the various entries.

Examples

d = (a = 0.153456453, b = 5.0, mode = "double")
savename(d; digits = 4) == "a=0.1535_b=5_mode=double"
savename("n", d) == "n_a=0.153_b=5_mode=double"
savename("n/", d) == "n/a=0.153_b=5_mode=double"
savename(d, "n") == "a=0.153_b=5_mode=double.n"
savename("data/n", d, "n") == "data/n_a=0.153_b=5_mode=double.n"
savename("n", d, "n"; connector = "-") == "n-a=0.153-b=5-mode=double.n"
savename(d, allowedtypes = (String,)) == "mode=double"

rick = (never = "gonna", give = "you", up = "!");
savename(rick) == "give=you_never=gonna_up=!" # keys are sorted!

DrWatson.@dict — Macro.

@dict vars...

Create a dictionary out of the given variables that has as keys the variable names and as values their values.

Examples

julia> ω = 5; χ = "test"; ζ = π/3;

julia> @dict ω χ ζ
Dict{Symbol,Any} with 3 entries:
  :ω => 5
  :χ => "test"
  :ζ => 1.0472

DrWatson.@strdict — Macro.

@strdict vars...

Same as @dict but the key type is String.

DrWatson.@ntuple — Macro.

@ntuple vars...

Create a NamedTuple out of the given variables that has as keys the variable names and as values their values.

Examples

julia> ω = 5; χ = "test"; ζ = 3.14;

julia> @ntuple ω χ ζ
(ω = 5, χ = "test", ζ = 3.14)

Notice that this naming scheme integrates perfectly with Parameters.jl.

Two convenience functions are also provided to easily switch between named tuples and dictionaries:

DrWatson.ntuple2dict — Function.

ntuple2dict(nt) -> dict

Convert a NamedTuple to a dictionary.

DrWatson.dict2ntuple — Function.

dict2ntuple(dict) -> ntuple

Convert a dictionary (with Symbol or String as key type) to a NamedTuple.

Customizing savename

You can customize savename for your own Types. For example you could make it so that it only uses some specific keys instead of all of them, only specific types, or you could make it access data in a different way (maybe even loading files!). You can even make it have a custom prefix!

To do that you may extend the following functions:

DrWatson.allaccess — Function.

allaccess(c)

Return all the keys c can be accessed using access. For dictionaries/named tuples this is keys(c), for everything else it is fieldnames(typeof(c)).

DrWatson.access — Function.

access(c, key)

Access c with given key. For AbstractDict this is getindex, for anything else it is getproperty.

access(c, keys...)

When given multiple keys, access is called recursively, i.e. access(c, key1, key2) = access(access(c, key1), key2) and so on. For example, if c is a NamedTuple then access(c, k1, k2) == ntuple.k1.k2.

Note

Please only extend the single key method when customizing access for your own Types.

DrWatson.default_allowed — Function.

default_allowed(c) = (Real, String, Symbol)

Return the (super-)Types that will be used as allowedtypes in savename or other similar functions.

DrWatson.default_prefix — Function.

default_prefix(c) = ""

Return the prefix that will be used by default in savename or other similar functions.

Tagging a run using Git

For reproducibility reasons (and also to not go insane when asking "HOW DID I GET THOSE RESUUUULTS") it is useful to "tag!" any simulation/result/process with the Git commit of the repository.

To this end there are two functions that can be used to ensure reproducibility:

DrWatson.current_commit — Function.

current_commit(path = projectdir()) -> commit

Return the current active commit id of the Git repository present in path, which by default is the project path. If the repository is dirty when this function is called the string will end with "_dirty".

See also tag!.

Examples

julia> current_commit()
"96df587e45b29e7a46348a3d780db1f85f41de04"

julia> current_commit(path_to_dirty_repo)
"3bf684c6a115e3dce484b7f200b66d3ced8b0832_dirty"

DrWatson.tag! — Function.

tag!(d::Dict, path = projectdir()) -> d

Tag d by adding an extra field commit which will have as value the current_commit of the repository at path (by default the project's path). Does not operate if a key commit already exists.

Notice that if String is not a subtype of the value type of d then a new dictionary is created and returned. Otherwise the operation is inplace (and the dictionary is returned again).

Examples

julia> d = Dict(:x => 3, :y => 4)
Dict{Symbol,Int64} with 2 entries:
  :y => 4
  :x => 3

julia> tag!(d)
Dict{Symbol,Any} with 3 entries:
  :y      => 4
  :commit => "96df587e45b29e7a46348a3d780db1f85f41de04"
  :x      => 3

Please notice that tag! will operate in place only when possible. If not possible then a new dictionary is returned. Also (importantly) these functions will never error as they are most commonly used when saving simulations and this could risk data not being saved.

Automatic Tagging during Saving

If you don't want to always call tag! before saving a file, you can just use the function tagsave:

DrWatson.tagsave — Function.

tagsave(file::String, d::Dict [, path = projectdir()])

First tag! dictionary d using the project in path, and then save d in file.

Produce or Load

produce_or_load is a function that very conveniently integrates with savename to either load a file if it exists, or if it doesn't to produce it, save it and then return it!

This saves you the effort of checking if a file exists and then loading, or then running some code and saving, or writing a bunch of if clauses in your code! produce_or_load really shines when used in interactive sessions where some results require a couple of minutes to complete.

DrWatson.produce_or_load — Function.

produce_or_load([prefix="",] c, f; kwargs...) -> file

Let s = savename(prefix, c, suffix). If a file named s exists then load it and return it.

If the file does not exist then call file = f(c), save file as s and then return the file.

To play well with BSON the function f should return a dictionary with Symbol as key type. The macro @dict can help with that.

Keywords

tag = false : Add the Git commit of the project in the saved file.

projectpath = projectdir() : Path to search for a Git repo.

suffix = "bson" : Used in savename.

kwargs... : All other keywords are propagated to savename.

See also savename and tag!.