Handling Simulations

Handling Simulations

This page discusses numerous tools that make life easier for handling simulations. Most (if not all) of these tools are also used in the examples demonstrated in the Real World Examples page. After reading the proper documentation here it might be worth it to have a look there as well!

Naming Schemes

A robust naming scheme allows you to create quick names for simulations, create lists of simulations, check existing simulations, etc. More importantly it allows you to easily read and write simulations using a consistent naming scheme.

DrWatson.savename — Function.

savename([prefix,], c [, suffix]; kwargs...)

Create a shorthand name, commonly used for saving a file, based on the parameters in the container c (Dict, NamedTuple or any other Julia composite type, e.g. created with Parameters.jl). If provided use the prefix and end the name with .suffix (i.e. you don't have to include the . in your suffix).

The function chains keys and values into a string of the form:

key1=val1_key2=val2_key3=val3

while the keys are always sorted alphabetically. If you provide the prefix/suffix the function will do:

prefix_key1=val1_key2=val2_key3=val3.suffix

assuming you chose the default connector, see below. Notice that if prefix can be any path and in addition if it ends as a path (/ or \) then the connector is ommited.

savename can be very conveniently combined with @dict or @ntuple.

Keywords

allowedtypes = (Real, String, Symbol) : Only values of type subtyping anything in allowedtypes are used in the name.

accesses = allaccess(c) : You can also specify which specific keys you want to use with the keyword accesses. By default this is all possible keys c can be accessed with, see allaccess.

digits = 3 : Floating point values are rounded to digits. In addition if the following holds:
round(val; digits = digits) == round(Int, val)
then the integer value is used in the name instead.

connector = "_" : string used to connect the various entries.

Examples

d = (a = 0.153456453, b = 5.0, mode = "double")
savename(d; digits = 4) == "a=0.1535_b=5_mode=double"
savename("n", d) == "n_a=0.153_b=5_mode=double"
savename("n/", d) == "n/a=0.153_b=5_mode=double"
savename(d, "n") == "a=0.153_b=5_mode=double.n"
savename("data/n", d, "n") == "data/n_a=0.153_b=5_mode=double.n"
savename("n", d, "n"; connector = "-") == "n-a=0.153-b=5-mode=double.n"
savename(d, allowedtypes = (String,)) == "mode=double"

rick = (never = "gonna", give = "you", up = "!");
savename(rick) == "give=you_never=gonna_up=!" # keys are sorted!

DrWatson.@dict — Macro.

@dict vars...

Create a dictionary out of the given variables that has as keys the variable names and as values their values.

Examples

julia> ω = 5; χ = "test"; ζ = π/3;

julia> @dict ω χ ζ
Dict{Symbol,Any} with 3 entries:
  :ω => 5
  :χ => "test"
  :ζ => 1.0472

DrWatson.@strdict — Macro.

@strdict vars...

Same as @dict but the key type is String.

DrWatson.@ntuple — Macro.

@ntuple vars...

Create a NamedTuple out of the given variables that has as keys the variable names and as values their values.

Examples

julia> ω = 5; χ = "test"; ζ = 3.14;

julia> @ntuple ω χ ζ
(ω = 5, χ = "test", ζ = 3.14)

Notice that this naming scheme integrates perfectly with Parameters.jl.

Two convenience functions are also provided to easily switch between named tuples and dictionaries:

ntupled2dict
dict2ntuple

Customizing savename

You can customize savename for your own Types. For example you could make it so that it only uses some specific keys instead of all of them, or you could make it access data in a different way (maybe even loading files!).

To do that you need to extend the following two functions:

DrWatson.allaccess — Function.

allaccess(c)

Return all the keys c can be accessed using access. For dictionaries/named tuples this is keys(c), for everything else it is fieldnames(typeof(c)).

DrWatson.access — Function.

access(c, key)

Access c with given key. For AbstractDict this is getindex, for anything else it is getproperty.

Tagging a run using Git

For reproducibility reasons (and also to not go insane when asking "HOW DID I GET THOSE RESUUUULTS") it is useful to "tag!" any simulation/result/process with the Git commit of the repository.

To this end there are two functions that can be used to ensure reproducibility:

DrWatson.current_commit — Function.

current_commit(path = projectdir()) -> commit

Return the current active commit id of the Git repository present in path, which by default is the project path. If the repository is dirty when this function is called the string will end with "_dirty".

See also tag!.

Examples

julia> current_commit()
"96df587e45b29e7a46348a3d780db1f85f41de04"

julia> current_commit(path_to_dirty_repo)
"3bf684c6a115e3dce484b7f200b66d3ced8b0832_dirty"

DrWatson.tag! — Function.

tag!(d::Dict, path = projectdir()) -> d

Tag d by adding an extra field commit which will have as value the current_commit of the repository at path (by default the project's path). Does not operate if a key commit already exists.

Notice that if String is not a subtype of the value type of d then a new dictionary is created and returned. Otherwise the operation is inplace (and the dictionary is returned again).

Examples

julia> d = Dict(:x => 3, :y => 4)
Dict{Symbol,Int64} with 2 entries:
  :y => 4
  :x => 3

julia> tag!(d)
Dict{Symbol,Any} with 3 entries:
  :y      => 4
  :commit => "96df587e45b29e7a46348a3d780db1f85f41de04"
  :x      => 3

Please notice that tag! will operate in place only when possible. If not possible then a new dictionary is returned. Also (importantly) these functions will never error as they are most commonly used when saving simulations and this could risk data not being saved.

Automatic Tagging during Saving

WIP. (adding the tag! functionality automatically with a save call)

Preparing Simulation Runs

It is very often the case that you want to run "batch simulations", i.e. just submit a bunch of different simulations, all using same algorithms and code but just different parameters. This scenario always requires the user to prepare a set of simulation parameter containers which are then passed into some kind of "main" function that starts the simulation.

To make the preparation part simpler we provide the following functionality:

DrWatson.dict_list — Function.

dict_list(c)

Expand the dictionary c into a vector of dictionaries. Each entry has a unique combination from the product of the Vector values of the dictionary while the non-Vector values are kept constant for all possibilities. The keys of the entries are the same.

Whether the values of c are iterable or not is of no concern; the function considers as "iterable" only subtypes of Vector.

Use the function dict_list_count to get an estimate of how many dictionaries will dict_list produce.

Examples

julia> c = Dict(:a => [1, 2], :b => 4);

julia> dict_list(c)
3-element Array{Dict{Symbol,Int64},1}:
 Dict(:a=>1,:b=>4)
 Dict(:a=>2,:b=>4)

julia> c[:model] = "linear"; c[:mode] = ["bi", "tri"];

julia> dict_list(c)
4-element Array{Dict{Symbol,Any},1}:
 Dict(:a=>1,:b=>4,:mode=>"bi",:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"bi",:model=>"linear")
 Dict(:a=>1,:b=>4,:mode=>"tri",:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"tri",:model=>"linear")

julia> c[:e] = [[1, 2], [3, 5]];

julia> dict_list(c)
8-element Array{Dict{Symbol,Any},1}:
 Dict(:a=>1,:b=>4,:mode=>"bi",:e=>[1, 2],:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"bi",:e=>[1, 2],:model=>"linear")
 Dict(:a=>1,:b=>4,:mode=>"tri",:e=>[1, 2],:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"tri",:e=>[1, 2],:model=>"linear")
 Dict(:a=>1,:b=>4,:mode=>"bi",:e=>[3, 5],:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"bi",:e=>[3, 5],:model=>"linear")
 Dict(:a=>1,:b=>4,:mode=>"tri",:e=>[3, 5],:model=>"linear")
 Dict(:a=>2,:b=>4,:mode=>"tri",:e=>[3, 5],:model=>"linear")

DrWatson.dict_list_count — Function.

dict_list_count(c) -> N

Return the number of dictionaries that will be created by calling dict_list(c).

Using the above function means that you can write your "preparation" step into a single dictionary and then let it automatically expand into many parameter containers. This keeps the code cleaner but also consistent, provided that it follows one rule: Anything that is a Vector has many parameters, otherwise it is one parameter. dict_list considers this true irrespectively of what the Vector contains. This allows users to use any iterable custom type as a single "parameter" of a simulation.

See the Real World Examples for a very convenient application!

Simulation Tables

WIP. (Adding simulation runs to a table/csv/dataframe)

Produce or Load

WIP. (loading a simulation or producing it if it doesn't exist)