Project Setup

Project Setup

Part of the functionality of DrWatson is creating and navigating through a project setup consistently. This works even if you move your project to a different location/computer and in addition the navigation process is identical across any project that uses DrWatson.

For this to work, you only need to follow these rules:

Your science project is also a Julia project defined by a Project.toml file.

You first activate this project environment before running any code. See Activating a Project for ways to do this.

You use the functions scriptdir, datadir, etc. from DrWatson (see Navigating a Project)

Default Project Setup

Here is the default project setup that DrWatson suggests (and assumes, for the functionality of this page):

|projectdir          <- Project's main folder. It is initialized as a Git
|                       repository with a reasonable .gitignore file.
│
├── _research        <- folder for WIP scripts, code, notes, comments,
|                       to-dos and anything in an alpha state
│
├── data             <- Folder that contains all data. **Immutable and add-only!**
│   ├── simulations  <- Data resulting directly from simulations
│   ├── exp_pro      <- Data from processing experiments
│   └── exp_raw      <- Raw experimental data
│
|
├── videos           <- self-explanatory
├── plots            <- self-explanatory
├── notebooks        <- Jupyter, Weave or any other mixed media notebooks.
│
├── papers           <- Scientific papers resulting from the project.
│
|
├── scripts          <- Various scripts, e.g. simulations, plotting, analysis,
|   |                   The scripts use the `src` folder for their base code.
│   └── intro.jl     <- simple file that uses DrWatson and uses its greeting.
|
├── src              <- Source code for use in this project. Contains functions,
|                       structures and modules that are used throughout
|                       the project and in multiple scripts.
│
|
├── README.md        <- Optional top-level README for anyone using this project.
├── .gitignore       <- by default ignores _research, data, plots, videos,
|                       notebooks and latex-compilation related files.
|
├── Manifest.toml    <- Contains full list of exact package versions used currently.
└── Project.toml     <- Main project file, allows activation and installation.
                        Includes DrWatson by default.

src vs scripts

Seems like src and scripts folders have pretty similar functionality. However there is a distinction between these two. You can follow these mental rules to know where to put file.jl:

If upon include("file.jl") there is anything being produced, be it data files, plots or even output to the console, then it should be in scripts.

If it is functionality used across multiple files or pipelines, it should be in src.

src should only contain files that define functions or types but not output anything. You can also organize src to be a Julia package, or contain multiple Julia packages.

Initializing a Project

To initialize a project as described in the Default Project Setup section, we provide the following function:

DrWatson.initialize_project — Function.

initialize_project(path [, name]; kwargs...)

Initialize a scientific project expected by DrWatson inside the given path. If its name is not given, it is assumed to be the folder's name.

The new project remains activated for you to immidiately add packages.

Keywords

readme = true : adds a README.md file.

authors = nothing : if a string or container of strings, adds the authors in the Project.toml file.

force = false : If the path is not empty then throw an error. If however force is true then recursively delete everything in the path and create the project.

git = true : Make the project a Git repository.

Activating a Project

This part of DrWatson's functionality requires you to have your scientific project (and as a consequence, the Julia project) activated. This can be done in multiple ways:

doing Pkg.activate("path/to/project") programmatically

using the startup flag --project path when starting Julia

by setting the JULIA_PROJECT environment variable

using the functions quickctivate and findproject offered by DrWatson.

We highly recommend the fourth approach. Here is how it works: the function quickactivate activates a project given some path by recursively searching the path and its parents for a valid Project.toml file. Typically you put this function in your script files like so:

using DrWatson # and any other package you use
quickactivate(@__DIR__)
# or
quickactivate(@__DIR__, "Best project in the WORLLDD")

where the second optional argument can assert if the activated project matches the name you provided, see below for more.

DrWatson.quickactivate — Function.

quickactivate(path [, name::String])

Activate the project found by findproject of the path. Optionally check if name is the same as the activated project's name. If it is not, throw an error.

This function is first activating the project and then checking if it matches the name.

DrWatson.findproject — Function.

findproject(path = pwd()) -> project_path

Recursively search path and its parents for a valid Julia project file. If it is found return its path, otherwise issue a warning and return nothing.

The function stops searching if it hits either the home directory or the root directory.

Navigating a Project

To be able to navigate the project consistently, DrWatson provides the following functions:

datadir() = projectdir()*"data/"
srcdir() = projectdir()*"src/"
plotsdir() = projectdir()*"plots/"
scriptdir() = projectdir()*"scripts/"

while as you can see all of them use projectdir:

DrWatson.projectdir — Function.

projectdir()

Return the directory of the currently active project. Ends with "/".

DrWatson.projectname — Function.

projectname()

Return the name of the currently active project.

In addition, all these functions end with / by default. This means that you can directly chain them with a file name. E.g. you could do

using DrWatson, FileIO
file = makesimulation()
FileIO.save(datadir()*"simulations/test.jld2", file)

Reproducibility

This project setup approach that DrWatson suggests has a very big side-benefit: it is fully reproducible firstly because it uses Julia's suggested project structure and secondly because the navigation only uses local directories.

If you send your entire project folder to a colleague, they only need to do:

julia> cd("path/to/project")
pkg> activate .
pkg> instantiate

All required packages and dependencies will be installed and then any script that was running in your computer will also be running in their computer in the same way!