Handling Simulations

Handling Simulations

Naming Schemes

A robust naming scheme allows you to create quick names for simulations, create lists of simulations, check existing simulations, etc.

DrWatson.savename — Function.

savename(d; kwargs...)

Create a shorthand name, commonly used for saving a file, based on the parameters in the container d (Dict, NamedTuple or any other Julia composite type, e.g. created with Parameters.jl).

The function chains keys and values into a string of the form:

key1=val1_key2=val2_key3=val3...

while the keys are always sorted alphabetically.

savename can be very conveniently combined with @dict or @ntuple.

Keywords

allowedtypes = (Real, String, Symbol)

Only values of type subtyping allowedtypes are used in the name.

accesses = allaccess(d)

You can also specify which specific keys you want to use with the keyword accesses. By default this is all possible keys d can be accessed with, see allaccess.

digits = 3

Floating point values are rounded to digits. In addition if the following holds:

round(val; digits = digits) == round(Int, val)

then the integer value is used in the name instead.

connector = "_" : string used to connect the various entries.

Examples

julia> d = (a = 0.153456453, b = 5.0, mode = "double")
(a = 0.153456453, b = 5.0, mode = "double")

julia> savename(d; digits = 4)
"a=0.1535_b=5_mode=double"

julia> savename(d, (String,))
"mode=double"

julia> rick = (never = "gonna", give = "you", up = "!");

julia> savename(rick) # keys are always sorted
  "give=you_never=gonna_up=!"

DrWatson.@dict — Macro.

@dict vars...

Create a dictionary out of the given variables that has as keys the variable names (as strings) and as values their values.

Examples

julia> ω = 5; χ = "test"; ζ = π/3;

julia> @dict ω χ ζ
Dict{String,Any} with 3 entries:
  "ω" => 5
  "χ" => "test"
  "ζ" => 1.0472

DrWatson.@ntuple — Macro.

@ntuple vars...

Create a NamedTuple out of the given variables that has as keys the variable names and as values their values.

Examples

julia> ω = 5; χ = "test"; ζ = 3.14;

julia> @ntuple ω χ ζ
(ω = 5, χ = "test", ζ = 3.14)

Notice that the naming scheme integrates perfectly with Parameters.jl.

Customizing savename

You can customize savename for your own Types. For example you could make it so that it only uses some specific keys instead of all of them, or you could make it access data in a different way (maybe even loading files!).

To do that you need to extend the following two functions:

DrWatson.allaccess — Function.

allaccess(d)

Return all the keys d can be accessed using access.

DrWatson.access — Function.

access(d, key)

Access d with given key. For AbstractDict this is getindex, for anything else it is getproperty.

Creating Run Tables

WIP. (Adding simulation runs to a table/csv/dataframe)

Produce or Load

WIP. (loading a simulation or producing it if it doesn't exist)